+++ forms-flow-api/src/formsflow_api/schemas/tasks.py
@@ -5,7 +5,7 @@ from marshmallow import EXCLUDE, fields
 from .base_schema import AuditDateTimeSchema
 
 
-class TaskOutcomeSchema(AuditDateTimeSchema):
+class TaskOutcomeConfigurationSchema(AuditDateTimeSchema):
     """This class manages task outcome configuration schema."""
 
     class Meta:  # pylint: disable=too-few-public-methods
@@ -13,8 +13,10 @@ class TaskOutcomeSchema(AuditDateTimeSchema):
 
         unknown = EXCLUDE
 
-    id = fields.Int()
+    id = fields.Int(dump_only=True)
     tenant = fields.Str(dump_only=True)
     task_id = fields.Str(data_key="taskId", required=True)
-    task_outcome = fields.List(fields.Dict(), data_key="taskOutcome", required=True)
+    task_name = fields.Str(data_key="taskName", required=True, allow_none=True)
+    task_transition_map = fields.Raw(data_key="taskTransitionMap", required=True)
+    transition_map_type = fields.Str(data_key="transitionMapType", required=True)
     created_by = fields.Str(data_key="createdBy", dump_only=True)